Corrected Indentation Error in conv_endian
The creation of a the hex_string variable was accidentally nested inside the conditional checking whether the _hex variable had an odd number of hex values. If the input was such that _hex had an even number of hex values, this would have caused a bug since the conditional checking for an odd number of hex values would've never been triggered, and thus hex_string would've never been assigned to the hex value of the input number. In that case an empty string (the initial value of hex_string) would've been returned. The only value being tested in the test suite resulted in an odd number of hex values in _hex so this bug wasn't caught by the test suite. I was just giving the code a quick once over and noticed that it was accidentally nested. We may want to think about testing a broader range of test values in our test suite. The conv_end function should be good to go now.